+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to the Audio Player project will be documented in this file.
 - [BUG] A song with title "0" cause scanner stopped to scan more music #601
 - fixed controls freezing when previous is clicked on the first track
 - reset album count between occ user scans
+- cover view no longer hides albums behind the song list
 
 ### Changed
 - Refactor SidebarController into service and mapper